requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.59.7
crewai @ git+https://github.com/joaomdmoura/crewAI.git
httpx>=0.27.2

# Telegram
python-telegram-bot>=21.10

# Twitter
tweepy>=4.14.0

# Vector Database
lancedb>=0.3.0

# LangChain
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-community>=0.0.13

# Document Processing
beautifulsoup4>=4.12.2
numpy>=1.24.3
requests>=2.31.0
pandas>=2.0.3

# Dependencies for document processing
tokenizers>=0.21.0
tiktoken>=0.8.0
pydantic>=2.4.2
tqdm>=4.65.0
chromadb>=0.5.23 